requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
google-generativeai
python-dotenv
fastapi
langserve
sse_starlette
uvicorn